Aligned some more variable names in wpa_network/wireless_network

+++ src/wpa-supplicant.h
@@ -37,8 +37,8 @@ struct ni_wpa_network {
 
 	int32_t			noise;
 	double			frequency;	/* in GHz */
-	int32_t			level;		/* 256 == 0dBm */
-	int32_t			quality;	/* n/70 */
+	double			level;		/* in dBm*/
+	double			quality;	/* n/70 */
 	unsigned int		max_bitrate;	/* in Bit/s, 1e6 based */
 	uint16_t		capabilities;
